Refactor: 화살표 관련 아이콘 통일 및 믹스인 추가
