+++ src/components/diaryForm/boardSection.scss
@@ -1,3 +1,5 @@
+@use '@/styles/mixins';
+
 .section_board {
   flex-grow: 1;
 
@@ -85,18 +87,14 @@
           }
         }
 
-        > .arrow_icon {
-          height: 6rem;
-          font-size: 1.6rem;
+        > .toggle {
+          @include mixins.arrow-icon('gray', 'bottom');
+          padding: 0;
+          width: 1.2rem;
+          height: 1.2rem;
 
-          display: flex;
-          align-items: center;
-          justify-content: flex-end;
-          flex-grow: 1;
-
-          > img {
-            width: 1rem;
-            height: 0.6rem;
+          &.open {
+            @include mixins.arrow-icon('gray', 'top');
           }
         }
       }
