+++ src/components/pageHeader/pageHeader.scss
@@ -37,8 +37,9 @@
     }
 
     > .back_btn {
-      width: 1.8rem;
-      height: 1.5rem;
+      padding: 0;
+      width: 2.4rem;
+      height: 2.4rem;
       background: url('@/assets/images/icons/back.png') no-repeat center /
         contain;
     }
