+++ src/pages/calendarPage/calendarPage.scss
@@ -1,4 +1,5 @@
 @use 'react-calendar/dist/Calendar.css';
+@use '@/styles/mixins';
 
 .hide {
   overflow: hidden;
@@ -44,18 +45,21 @@
       }
 
       .react-calendar__navigation__prev-button {
+        @include mixins.arrow-icon('blackLight', 'left');
+        background-color: unset;
         padding: 0;
+        min-width: unset;
         width: 1.8rem;
         height: 1.8rem;
-        background: url('@/assets/images/icons/calendar_prev.png') no-repeat
-          left center / contain;
       }
+
       .react-calendar__navigation__next-button {
+        @include mixins.arrow-icon('blackLight', 'right');
+        background-color: unset;
         padding: 0;
+        min-width: unset;
         width: 1.8rem;
         height: 1.8rem;
-        background: url('@/assets/images/icons/calendar_next.png') no-repeat
-          right center / contain;
       }
 
       .react-calendar__navigation__label {
