+++ src/pages/myPage/myInfoPage/myInfoPage.scss
@@ -1,16 +1,3 @@
-.sub_header {
-  .back_btn {
-    position: absolute;
-    left: 2rem;
-    top: 50%;
-    transform: translateY(-50%);
-    width: 2.4rem;
-    height: 2.4rem;
-    background: url('../../assets/images/icons/back.png') no-repeat center /
-      contain;
-  }
-}
-
 .my_info_page {
   .my_info_container {
     height: 100vh;
@@ -35,11 +22,25 @@
         position: absolute;
         right: -0.4rem;
         bottom: 0;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        background-color: white;
+        border: 1px solid #9b9fb1;
+        border-radius: 50%;
         width: 2.4rem;
         height: 2.4rem;
-        background: url('../../assets/images/icons/icon_edit.png') no-repeat
-          center / contain;
         filter: drop-shadow(0px 4px 10px rgba(0, 0, 0, 0.1));
+        cursor: pointer;
+
+        &::after {
+          content: '';
+          display: inline-block;
+          background: url('@/assets/images/icons/edit.png') no-repeat center /
+            contain;
+          width: 1.4rem;
+          height: 1.4rem;
+        }
       }
 
       > input {
